suuuuAdd

+++ frontend/src/pages/ConfirmAdvertisement.js
@@ -24,13 +24,26 @@ const ConfirmAdvertisement = () => {
 
     }, []);
 
+    // const handleDelete = async (id) => {
+    //     try {
+    //         await axios.delete(`http://localhost:9000/ads/${id}`);
+    //         setAds(ads.filter((add) => add._id !== id));
+    //         alert('Ad deleted successfully');
+    //     } catch (error) {
+    //         alert('Failed to delete ad');
+    //     }
+    // };
     const handleDelete = async (id) => {
-        try {
-            await axios.delete(`http://localhost:9000/ads/${id}`);
-            setAds(ads.filter((add) => add._id !== id));
-            alert('Ad deleted successfully');
-        } catch (error) {
-            alert('Failed to delete ad');
+        if (window.confirm("Are you sure you want to delete this customer?")) {
+            try {
+                await axios.delete(`http://localhost:9000/ads/${id}`);
+                setAds(ads.filter((ad) => ad._id !== id));
+                alert('Add deleted successfully');
+            } catch (error) {
+                alert('Failed to delete add');
+            }
+        } else {
+            alert('Deletion cancelled.');
         }
     };
 
@@ -67,7 +80,7 @@ const ConfirmAdvertisement = () => {
                    &nbsp;Confirm
                 </a>
                 &nbsp;
-                <button className= "btn btn-danger" onClick={() => handleDelete(ads._id)}>Delete</button>
+                <button className= "btn btn-danger" onClick={() => handleDelete(ad._id)}>Delete</button>
                 </div>
               </td>
                         </tr>
